# Case when performance improvement: reduce the `copy_if_else` [databricks]