--- a/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuColumnVectorUtils.scala
+++ b/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuColumnVectorUtils.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2022-2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,10 +18,11 @@ package com.nvidia.spark.rapids
 
 import java.lang.reflect.Method
 
-import ai.rapids.cudf.Table
+import ai.rapids.cudf.{ColumnVector => CudfCV, Table}
 
-import org.apache.spark.sql.types.DataType
+import org.apache.spark.sql.types._
 import org.apache.spark.sql.vectorized.ColumnVector
+import org.apache.spark.unsafe.types.UTF8String
 
 object GpuColumnVectorUtils {
   lazy val extractHostColumnsMethod: Method = ShimLoader.loadGpuColumnVector()
@@ -38,4 +39,54 @@ object GpuColumnVectorUtils {
     columnVectors.asInstanceOf[Array[ColumnVector]]
   }
 
+  def isCaseWhenFusionSupportedType(dataType: DataType): Boolean = {
+    dataType match {
+      case BooleanType => true
+      case ByteType => true
+      case ShortType => true
+      case IntegerType => true
+      case LongType => true
+      case FloatType => true
+      case DoubleType => true
+      case StringType => true
+      case DecimalType() => true
+      case _ => false
+    }
+  }
+
+  def createFromScalarList(scalars: Seq[GpuScalar]): CudfCV = {
+    scalars.head.dataType match {
+      case BooleanType =>
+        val booleans = scalars.map(s => s.getValue.asInstanceOf[java.lang.Boolean])
+        CudfCV.fromBoxedBooleans(booleans: _*)
+      case ByteType =>
+        val bytes = scalars.map(s => s.getValue.asInstanceOf[java.lang.Byte])
+        CudfCV.fromBoxedBytes(bytes: _*)
+      case ShortType =>
+        val shorts = scalars.map(s => s.getValue.asInstanceOf[java.lang.Short])
+        CudfCV.fromBoxedShorts(shorts: _*)
+      case IntegerType =>
+        val ints = scalars.map(s => s.getValue.asInstanceOf[java.lang.Integer])
+        CudfCV.fromBoxedInts(ints: _*)
+      case LongType =>
+        val longs = scalars.map(s => s.getValue.asInstanceOf[java.lang.Long])
+        CudfCV.fromBoxedLongs(longs: _*)
+      case FloatType =>
+        val floats = scalars.map(s => s.getValue.asInstanceOf[java.lang.Float])
+        CudfCV.fromBoxedFloats(floats: _*)
+      case DoubleType =>
+        val doubles = scalars.map(s => s.getValue.asInstanceOf[java.lang.Double])
+        CudfCV.fromBoxedDoubles(doubles: _*)
+      case StringType =>
+        val utf8Bytes = scalars.map(s => s.getValue.asInstanceOf[UTF8String].getBytes)
+        CudfCV.fromUTF8Strings(utf8Bytes: _*)
+      case DecimalType() =>
+        val utf8Bytes = scalars.map(s => s.getValue.asInstanceOf[Decimal].toBigDecimal
+            .asInstanceOf[java.math.BigDecimal])
+        CudfCV.fromDecimals(utf8Bytes: _*)
+      case _ =>
+        throw new UnsupportedOperationException(s"Creating column vector from a GpuScalar list" +
+            s" is not supported for type ${scalars.head.dataType}.")
+    }
+  }
 }