--- a/sql-plugin/src/main/scala/com/nvidia/spark/rapids/conditionalExpressions.scala
+++ b/sql-plugin/src/main/scala/com/nvidia/spark/rapids/conditionalExpressions.scala
@@ -318,7 +318,8 @@ case class GpuIf(
 case class GpuCaseWhen(
     branches: Seq[(Expression, Expression)],
     elseValue: Option[Expression] = None,
-    caseWhenFuseEnabled: Boolean = true) extends GpuConditionalExpression with Serializable with Logging {
+    caseWhenFuseEnabled: Boolean = true)
+    extends GpuConditionalExpression with Serializable {
 
   import GpuExpressionWithSideEffectUtils._
 
@@ -363,27 +364,31 @@ case class GpuCaseWhen(
     if (branchesWithSideEffects) {
       columnarEvalWithSideEffects(batch)
     } else {
-      if (caseWhenFuseEnabled &&
+      if (caseWhenFuseEnabled && branches.size > 2 &&
           inputTypesForMerging.head == StringType &&
         (branches.map(_._2) ++ elseValue).forall(_.isInstanceOf[GpuLiteral])
       ) {
-        // return type is string type; all the then and else exprs are Scalars
-        // avoid to use multiple `computeIfElse`s which will create multiple temp columns
-        logWarning("==================== Running case with experimental =========== ")
+        // when branches size > 2;
+        // return type is string type;
+        // all the then and else exprs are Scalars.
+        // Avoid to use multiple `computeIfElse`s which will create multiple temp columns
 
+        // 1. select first true index from bool columns
         val whenBoolCols = branches.safeMap(_._1.columnarEval(batch).getBase).toArray
         val firstTrueIndex: ColumnVector = withResource(whenBoolCols) { _ =>
           CaseWhen.selectFirstTrueIndex(whenBoolCols)
         }
 
         withResource(firstTrueIndex) { _ =>
-          val thenElseScalars = (branches.map(_._2) ++ elseValue).map(_.columnarEvalAny(batch).asInstanceOf[GpuScalar])
+          val thenElseScalars = (branches.map(_._2) ++ elseValue).map(_.columnarEvalAny(batch)
+              .asInstanceOf[GpuScalar])
           withResource(thenElseScalars) { _ =>
-            // generate a column to store all scalars
-            val scalarsBytes = thenElseScalars.map(ret => ret.getValue.asInstanceOf[UTF8String].getBytes)
+            // 2. generate a column to store all scalars
+            val scalarsBytes = thenElseScalars.map(ret => ret.getValue
+                .asInstanceOf[UTF8String].getBytes)
             val scalarCol = ColumnVector.fromUTF8Strings(scalarsBytes: _*)
             withResource(scalarCol) { _ =>
-              // execute final select
+              // 3. execute final select
               val finalRet = CaseWhen.selectFromIndex(scalarCol, firstTrueIndex)
               // return final column vector
               GpuColumnVector.from(finalRet, dataType)