contributes to [#2084](https://github.com/NVIDIA/spark-rapids-jni/issues/2084)

### Current case when logic: [link](https://github.com/NVIDIA/spark-rapids/blob/branch-24.06/sql-plugin/src/main/scala/com/nvidia/spark/rapids/conditionalExpressions.scala#L362-L370)
Iteratively invoke copy_if_else to merge the tail 2 branches.
This will create lots of temp string columns via `copy_if_else` which will introduce more spend time.

```
      val elseRet = elseValue
        .map(_.columnarEvalAny(batch))
        .getOrElse(GpuScalar(null, branches.last._2.dataType))
      val any = branches.foldRight[Any](elseRet) {
        case ((predicateExpr, trueExpr), falseRet) =>
          computeIfElse(batch, predicateExpr, trueExpr, falseRet)
      }
```
### new logic
For the `case when` with multiple branches and the `then` values are all string scalars.
e.g.:

```
CASE 
  WHEN expr1 THEN 's1'
  WHEN expr2 THEN 's2'
  WHEN expr2 THEN 's3' 
   ...
  ELSE 's-default'
END
```

This PR removed the `copy_if_else`s.
Steps:
- First evaluate all the `expr1` expr2...., and get bool columns.
- Generate a column for all the scalars.
- Then select from bool table(bool columns), get all the expected index of scalars for each row, and generate a int column contains scalar indexes.
- Then select scalars from scalar column according to the selecar indexes, and generate a String column.

For more details, refer to JNI PR: https://github.com/NVIDIA/spark-rapids-jni/pull/2079
Depending on PR: https://github.com/NVIDIA/spark-rapids-jni/pull/2079

Signed-off-by: Chong Gao <res_life@163.com>