@res-life 2024-06-03
Perf test result:

|case when branches|legacy min time|new impl min time|speedup|
|:-|:-|:-|:-|
|3 | 147 |129 |0.14x|
|10 | 197 |298 |0.51x|

Test steps:
- create 5 parquet files with 10 bool columns in it.
Each file contains 10,000,000 rows:

```
+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+
|  b01|  b02|  b03|  b04|  b05|  b06|  b07|  b08|  b09|  b10|
+-----+-----+-----+-----+-----+-----+-----+-----+-----+-----+
|false| true|false|false| true|false|false| true|false|false|
| true|false| true|false|false|false| true| true|false|false|
|false| true| true| true| true|false| true|false|false| true|
| true| true|false|false| true|false|false|false| true|false|
|false| true|false| true|false|false|false| true| true| true|
| true|false|false| true|false|false|false| true| true| true|
|false|false|false|false| true| true|false|false|false| true|
```
- SQL
10 branches:
```
"""
select
  sum(octet_length(
    CASE WHEN b01 THEN 'go-http-client'
         WHEN b02 THEN 'libcurlOT'
         WHEN b03 THEN 'libcurl'
         WHEN b04 THEN 'wspcdn'
         WHEN b05 THEN 'byte-pcdn-vod'
         WHEN b06 THEN 'byte-pcdn-live'
         WHEN b07 THEN 'byte-pcdn/vod-go'
         WHEN b08 THEN 'byte-pcdn/vod-bs'
         WHEN b09 THEN 'kcg-cache'
         WHEN b10 THEN 'byte-pcdn-vod'
         ELSE 'unknown'
    END
  ))
from tab
"""
```

3 branches:
```
"""
select
  sum(octet_length(
    CASE WHEN b01 THEN 'go-http-client'
         WHEN b02 THEN 'libcurlOT'
         WHEN b03 THEN 'libcurl'
         ELSE 'unknown'
    END
  ))
from tab
"""
```