@res-life 2024-06-04
This PR only improves the case that all `when` expressions are scalars.
For the general case that `when` expressions are not all scalars, we may file a follow-up to imporve.
IMO, for the case that return type is string,  the method in this PR will help; For other type, it may not help much.

This PR only improves the non side effect code.