@res-life 2024-07-04
Perf tests for other types:
|type|when branches num|new MS(enable optimazation) (2nd,3rd,4th) |baseline MS(disable optimazation) (2nd,3rd,4th) | end to end sppedup (new MS/baseline MS)|
|:-|:-|:-|:-|:-|
|bool | 3    |133, 124, 121 | 144, 139, 139|1.11640211640212 |
|int | 3     |159,148, 146  | 160, 145, 157|1.01986754966887 |
|byte | 3    |127, 140, 130 | 151, 147, 152|1.13350125944584 |
|float | 3   |120, 133, 131 | 153, 143, 149|1.15885416666667 |
|decimal32|3 |126, 135, 122 | 140, 141, 133|1.08093994778068 |
|bool | 10   |172, 199, 205 | 267, 280, 280|1.43576388888889 |
|int | 10    |215,215, 207  | 263, 273, 266|1.25902668759812 |
|byte | 10   |210, 180, 201 | 261, 287, 289|1.41624365482234 |
|float | 10  |194, 177, 209 | 281, 262, 275|1.41034482758621 |
|decimal32|10|177, 194, 200 | 265, 268, 270|1.40630472854641 |