@res-life 2024-07-11
Premerge failure reason: 
[parquet_test.py::test_buckets failed due to SparkOutOfMemoryError](https://github.com/NVIDIA/spark-rapids/issues/10940#top)